Add max length and max duration parameters
